+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to the
 
 ## Unreleased
 
+- Changed default commit range from `HEAD` to `HEAD~1..HEAD`
+
 ## 0.1.0 - 2025-10-07
 
 Initial release.