+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to the
 
 - Changed default commit range from `HEAD` to `HEAD~1..HEAD`.
 - Changed return code to failure if any checks failed during the run.
+- Changed config file resolution to be relative to repository under check.
 
 ## 0.1.0 - 2025-10-07